Refactor (frontend): 파일 이름 변경 및 경로 수정
